audio-extraction-service/requirements.txt
Flask>=2.0.0
google-cloud-storage>=1.40.0
gunicorn>=20.0.0
